backend/requirements/common.txt
asgiref==3.5.2
Django==4.1.2
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.14.0
drf-extra-fields==3.4.1
Pillow==9.2.0
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
